+++ NEWS.md
@@ -28,6 +28,11 @@
 - `XPA.list` can be called with a predicate function to filter the servers based on their
   `XPA.AccessPoint`: `XPA.list(f)` is the same as `filter(f, XPA.list()))`.
 
+- `XPA.find` can also be called with a predicate function and may interact with the user
+  to select a single access-point.
+
+- Previous methods for `XPA.list` and `XPA.find` have been deprecated.
+
 - Private abstract type `XPA.Handle` removed. Use `Union{XPA.Client,XPA.Server}` instead.